backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
python-dateutil==2.8.2
pytz==2023.3